Format and make a new int test suite
