+++ kotlin-client/src/integrationTest/kotlin/io/provenance/bilateral/contract/SearchIntTest.kt
@@ -1,8 +1,6 @@
 package io.provenance.bilateral.contract
 
 import cosmos.tx.v1beta1.ServiceOuterClass.BroadcastMode
-import io.provenance.bilateral.execute.CancelAsk
-import io.provenance.bilateral.execute.CancelBid
 import io.provenance.bilateral.execute.CreateAsk
 import io.provenance.bilateral.execute.CreateBid
 import io.provenance.bilateral.models.AttributeRequirement
